ssltest: take into account the absence of common ciphers
During certificate trust test there can be a situation when client does
not accept any of the suggested ciphers. In this case it will refuce TLS
connection. However, it does not mean that it trusts the certificate.

This commit changes the behavior and the tool returns UNDEFINED test
result in such situation. See also Github issue #18.